finished agreements backend
