+++ nextjs_frontend/src/models/Acordos.tsx
@@ -1,23 +1,32 @@
 import mongoose from "mongoose";
 
-export type StatusType =
-  | "Aguardando inadimplente"
-  | "Conversa iniciada"
-  | "Valor reserva alcançado"
-  | "Negociação concluída";
+export enum StatusType {
+  "Aguardando inadimplente",
+  "Conversa iniciada",
+  "Primeira proposta realizada",
+  "Segunda proposta realizada",
+  "Terceira proposta realizada",
+  "Aguardando confirmação da proposta",
+  "Acordo realizado",
+  "Acordo rejeitado",
+}
+
+export interface Proposta {
+  entrada: number;
+  valorParcela: number;
+  qtdParcelas: number;
+}
 
 export interface Acordo {
   usuarioEmail: string;
   cpfDevedor: string;
   dataAcordo?: Date;
+  status: StatusType;
+
   entrada: number;
   valorTotal: number;
   qtdParcelas: number;
-}
-
-export interface AcordoIdentificado extends Acordo {
-  nomeDevedor: string;
-  nomeCondominio: string;
+  historicoValores: Proposta[];
 }
 
 const AcordoSchema = new mongoose.Schema({
@@ -33,9 +42,13 @@ const AcordoSchema = new mongoose.Schema({
     type: Date,
     default: Date.now,
   },
+  status: String,
   entrada: Number,
   valorTotal: Number,
   qtdParcelas: Number,
+  historicoValores: [
+    { entrada: Number, valorParcela: Number, qtdParcelas: Number },
+  ],
 });
 
 export default mongoose.models.Acordos ||
